+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Perform frustum culling for voxel meshes (#1183, **@mkuritsu**).
 - Infinite terrain profiling sample (#1487, **@tomas7770**).
 - Added inertia for voxel shape (**@fallenatlas**).
+- Added a method to Query to count the number of matches (#1451, **@GalaxyCrush**).
 
 ### Changed
 
@@ -25,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Deferred shading missing padding leading to black screen on web builds (#1476, **@RiscadoA**).
+- Crashes when adding shadow casters (#1471, **@GalaxyCrush**).
 
 ## [v0.6.0] - 2025-02-10